requirements.txt
streamlit
google-auth
google-auth-oauthlib
google-auth-httplib2
google-api-python-client
langchain
langchain-community
chromadb
pypdf
ollama
python-dotenv
